+++ docs/tutorial/shorthands_store.js
@@ -140,7 +140,7 @@ Promise.all([dataLoaded, mdChartLoaded]).then((results) => {
 					return chart.animate({
 						style: {
 							'plot.xAxis.label.fontSize': '150%',
-							'title.backgroundColor': '#A0A0A0'
+							'plot.backgroundColor': '#A0A0A0'
 						}
 					})
 				}