+++ src/chart/main/style.h
@@ -106,9 +106,7 @@ struct Font
 	[[nodiscard]] std::string calculatedFamily() const
 	{
 		if (fontFamily.has_value())
-			if (auto &&ff =
-			        fontFamily->get_or_first(::Anim::first).value;
-			    !ff.empty())
+			if (auto &&ff = fontFamily->values[0].value; !ff.empty())
 				return ff;
 
 		if (fontParent) return fontParent->calculatedFamily();
@@ -363,9 +361,9 @@ struct PlotParams
 
 struct Plot : Padding, Box, PlotParams
 {
-	[[nodiscard]] const Axis &getAxis(Gen::ChannelId id) const
+	[[nodiscard]] const Axis &getAxis(Gen::AxisId id) const
 	{
-		return id == Gen::ChannelId::x ? xAxis : yAxis;
+		return id == Gen::AxisId::x ? xAxis : yAxis;
 	}
 };