+++ src/chart/options/channel.cpp
@@ -16,9 +16,22 @@
 namespace Vizzu::Gen
 {
 
-bool isAxis(ChannelId type)
+std::optional<AxisId> asAxis(ChannelId type)
 {
-	return type == ChannelId::x || type == ChannelId::y;
+	return type == ChannelId::x || type == ChannelId::y
+	         ? std::make_optional(static_cast<AxisId>(
+	             static_cast<ChannelIdType>(type)))
+	         : std::nullopt;
+}
+
+ChannelId asChannel(AxisId type)
+{
+	return static_cast<ChannelId>(static_cast<ChannelIdType>(type));
+}
+
+bool operator==(const AxisId &axis, const ChannelId &channel)
+{
+	return asChannel(axis) == channel;
 }
 
 Channel Channel::makeChannel(Type id)