+++ src/chart/rendering/drawinterlacing.cpp
@@ -39,8 +39,7 @@ void DrawInterlacing::drawTexts() const
 
 void DrawInterlacing::draw(bool horizontal, bool text) const
 {
-	auto axisIndex =
-	    horizontal ? Gen::ChannelId::y : Gen::ChannelId::x;
+	auto axisIndex = horizontal ? Gen::AxisId::y : Gen::AxisId::x;
 
 	auto interlacingColor =
 	    *rootStyle.plot.getAxis(axisIndex).interlacing.color;
@@ -111,8 +110,7 @@ void DrawInterlacing::draw(
 	const auto &enabled =
 	    horizontal ? plot->guides.y : plot->guides.x;
 
-	auto axisIndex =
-	    horizontal ? Gen::ChannelId::y : Gen::ChannelId::x;
+	auto axisIndex = horizontal ? Gen::AxisId::y : Gen::AxisId::x;
 
 	const auto &axisStyle = rootStyle.plot.getAxis(axisIndex);
 
@@ -148,7 +146,8 @@ void DrawInterlacing::draw(
 		auto iMin =
 		    axisBottom > 0
 		        ? std::floor(-axis.origo() / (2 * stripWidth)) * 2
-		        : std::round(axis.range.getMin() - stepSize);
+		        : std::round(
+		            (axis.range.getMin() - stepSize) / stepSize);
 
 		auto interlaceCount = 0U;
 		const auto maxInterlaceCount = 1000U;
@@ -251,8 +250,7 @@ void DrawInterlacing::drawDataLabel(
     const ::Anim::String &unit,
     double alpha) const
 {
-	auto axisIndex =
-	    horizontal ? Gen::ChannelId::y : Gen::ChannelId::x;
+	auto axisIndex = horizontal ? Gen::AxisId::y : Gen::AxisId::x;
 	const auto &labelStyle = rootStyle.plot.getAxis(axisIndex).label;
 
 	auto drawLabel = OrientedLabel{{ctx()}};
@@ -308,8 +306,7 @@ void DrawInterlacing::drawSticks(double tickIntensity,
     bool horizontal,
     const Geom::Point &tickPos) const
 {
-	auto axisIndex =
-	    horizontal ? Gen::ChannelId::y : Gen::ChannelId::x;
+	auto axisIndex = horizontal ? Gen::AxisId::y : Gen::AxisId::x;
 	const auto &axisStyle = rootStyle.plot.getAxis(axisIndex);
 	const auto &tickStyle = axisStyle.ticks;