+++ src/chart/rendering/drawmarkerinfo.cpp
@@ -195,8 +195,7 @@ void DrawMarkerInfo::draw(Gfx::ICanvas &canvas,
     const Geom::Rect &boundary) const
 {
 	for (const auto &info : plot->getMarkersInfo()) {
-		auto &&[cnt1, weight1] =
-		    info.second.get_or_first(::Anim::first);
+		auto &&[cnt1, weight1] = info.second.values[0];
 		if (!info.second.interpolates() && cnt1) {
 			MarkerDC dc(*this, canvas, boundary, cnt1);
 			dc.draw(weight1);