+++ src/chart/rendering/markers/connectingmarker.cpp
@@ -29,13 +29,36 @@ ConnectingMarker::ConnectingMarker(const DrawingContext &ctx,
 	auto isLine = type == Gen::ShapeType::line;
 	auto isArea = type == Gen::ShapeType::area;
 
+	auto needConnection =
+	    !ctx.getOptions().geometry.interpolates()
+	    || isConnecting(
+	        ctx.getOptions().geometry.get_or_first(lineIndex).value);
+
+	auto otherNeedConnection = isConnecting(
+	    ctx.getOptions()
+	        .geometry
+	        .get_or_first(lineIndex == ::Anim::first ? ::Anim::second
+	                                                 : ::Anim::first)
+	        .value);
+
 	auto polar = ctx.getOptions().coordSystem.factor<Math::FuzzyBool>(
 	    Gen::CoordSystem::polar);
 	auto horizontal =
 	    ctx.getOptions().orientation.factor<Math::FuzzyBool>(
 	        Gen::Orientation::horizontal);
 
-	linear = !polar || horizontal;
+	auto &&isHorizontal =
+	    ctx.getOptions().orientation.get_or_first(lineIndex).value
+	    == Gen::Orientation::horizontal;
+
+	linear = !polar || horizontal
+	      || Math::FuzzyBool::And(
+	          !isHorizontal
+	              && ctx.plot->axises.at(Gen::Channel::Type::x)
+	                     .dimension.enabled,
+	          ctx.plot->axises.at(Gen::Channel::Type::x)
+	              .measure.enabled.factor(false),
+	          !horizontal);
 
 	lineWidth[0] = lineWidth[1] = 0;
 
@@ -49,7 +72,8 @@ ConnectingMarker::ConnectingMarker(const DrawingContext &ctx,
 	if (prev) {
 		enabled = labelEnabled && prev->enabled;
 		connected = Math::FuzzyBool::And(enabled,
-		    marker.prevMainMarker.get_or_first(lineIndex).weight);
+		    marker.prevMainMarker.get_or_first(lineIndex).weight,
+		    needConnection);
 		if (auto &&pc =
 		        marker.polarConnection.get_or_first(lineIndex);
 		    pc.value) {
@@ -59,6 +83,9 @@ ConnectingMarker::ConnectingMarker(const DrawingContext &ctx,
 			      || Math::FuzzyBool::Or(polar, pc.weight).more();
 			connected = connected && newPolar && horizontal;
 			enabled = enabled && newPolar && horizontal;
+
+			if (polar != false && prev != &marker)
+				linear = linear || Math::FuzzyBool{isHorizontal};
 		}
 	}
 	else {
@@ -92,13 +119,13 @@ ConnectingMarker::ConnectingMarker(const DrawingContext &ctx,
 			auto prevSpacing = prev->spacing * prev->size / 2;
 			auto prevPos = prev->position;
 
-			if (polar != false
-			    && ctx.getOptions()
-			               .orientation.get_or_first(lineIndex)
-			               .value
-			           == Gen::Orientation::horizontal
-			    && prev != &marker && prevPos.x >= 1) {
-				prevPos.x -= 1;
+			if (polar != false && isHorizontal && prev != &marker) {
+				if (!otherNeedConnection
+				    && marker.polarConnection.get_or_first(lineIndex)
+				           .value)
+					prevPos.x = 0;
+				else if (prevPos.x >= 1)
+					prevPos.x -= 1;
 			}
 
 			points[3] = prevPos - prevSpacing;